Clean up and sort player files
