+++ src/components/karaoke/player/Player.jsx
@@ -1,16 +1,18 @@
 import classNames from 'classnames'
-import dayjs from 'dayjs'
-import relativeTime from 'dayjs/plugin/relativeTime'
 import PropTypes from 'prop-types'
 import queryString from 'query-string'
 import { Component } from 'react'
-import { connect, useSelector } from 'react-redux'
-import { Link } from 'react-router'
+import { connect } from 'react-redux'
 
 import { sendPlayerCommand } from 'actions/playlist'
+import { Carousel } from 'components/generics/Carousel'
+import {
+  CarouselEntryCurrentSong,
+  CarouselEntryNextSong,
+  CarouselEntryStats,
+} from 'components/karaoke/player/Carousel'
 import ManageButton from 'components/karaoke/player/ManageButton'
 import PlayerNotification from 'components/karaoke/player/Notification'
-import PlaylistEntryWidget from 'components/playlist/widgets/PlaylistEntry'
 import { IsPlaylistManagerOrOwner } from 'permissions/Playlist'
 import {
   alterationResponsePropType,
@@ -21,9 +23,6 @@ import { playerErrorsDigestStatePropType } from 'reducers/playlistDigest'
 import { userPropType } from 'serverPropTypes/users'
 import { withNavigate } from 'thirdpartyExtensions/ReactRouterDom'
 import { CSSTransitionLazy } from 'thirdpartyExtensions/ReactTransitionGroup'
-import { formatDate, formatDuration, isDisplayable } from 'utils'
-
-dayjs.extend(relativeTime)
 
 class Player extends Component {
   static propTypes = {
@@ -243,225 +242,3 @@ Player = withNavigate(
 )
 
 export default Player
-
-function Carousel({ children, className }) {
-  return (
-    <div className={classNames('carousel', className)}>
-      <ul className="viewport">{children}</ul>
-    </div>
-  )
-}
-
-Carousel.propTypes = {
-  children: PropTypes.node,
-  className: PropTypes.string,
-}
-
-function CarouselEntry({ children, className, jumbo }) {
-  return (
-    <li className="carousel-entry">
-      {isDisplayable(jumbo) && <div className="jumbo">{jumbo}</div>}
-      <div className="content" tabIndex="0">
-        <div className={className}>{children}</div>
-      </div>
-    </li>
-  )
-}
-
-CarouselEntry.propTypes = {
-  children: PropTypes.node,
-  className: PropTypes.string,
-  jumbo: PropTypes.string,
-}
-
-function CarouselEntryCurrentSong() {
-  const { data: playerStatus } = useSelector(
-    (state) => state.playlist.playerStatus
-  )
-  const { playlist_entry: entry, timing } = playerStatus
-
-  if (!entry) {
-    return null
-  }
-
-  return (
-    <CarouselEntry className="current-song">
-      <Link
-        className="to-song"
-        to={{
-          pathname: '/library/song',
-          search: queryString.stringify({
-            query: `title:""${entry.song.title}""`,
-            expanded: entry.song.id,
-          }),
-        }}
-      >
-        <PlaylistEntryWidget
-          entry={entry}
-          noRelativeDate
-          songProps={{ noRelations: false }}
-        />
-      </Link>
-      <div className="timing">
-        <div className="current">{formatDuration(timing)}</div>
-        <div className="duration">{formatDuration(entry.song.duration)}</div>
-      </div>
-    </CarouselEntry>
-  )
-}
-
-function CarouselEntryNextSong() {
-  const { playlistEntries } = useSelector(
-    (state) => state.playlist.digest.entries.data
-  )
-  const entry = playlistEntries.find((e) => e.will_play)
-
-  if (!entry) {
-    return null
-  }
-
-  return (
-    <CarouselEntry jumbo="Next" className="next-song">
-      <Link
-        to={{
-          pathname: '/library/song',
-          search: queryString.stringify({
-            query: `title:""${entry.song.title}""`,
-            expanded: entry.song.id,
-          }),
-        }}
-      >
-        <PlaylistEntryWidget entry={entry} />
-      </Link>
-    </CarouselEntry>
-  )
-}
-
-function CarouselEntryStats() {
-  const { playlistEntries, dateEnd } = useSelector(
-    (state) => state.playlist.digest.entries.data
-  )
-  const { data: playerStatus } = useSelector(
-    (state) => state.playlist.playerStatus
-  )
-  const { date_stop: karaokeDateStop } = useSelector(
-    (state) => state.playlist.karaoke.data
-  )
-  const countPlayed = playlistEntries.filter((e) => e.was_played).length
-  const countQueueing = playlistEntries.filter((e) => e.will_play).length
-
-  /**
-   * Played songs
-   */
-
-  let played
-  switch (countPlayed) {
-    case 0:
-      played = <li>The karaoke has just started!</li>
-      break
-    case 1:
-      played = <li>One song played, keep going!</li>
-      break
-    default:
-      played = (
-        <li>
-          <q>{countPlayed}</q> songs played so far
-        </li>
-      )
-  }
-
-  /**
-   * Queuing songs
-   */
-
-  let queuing
-  switch (countQueueing) {
-    case 0:
-      queuing = <li>No songs queued in playlist yet</li>
-      break
-    case 1:
-      queuing = <li>One song queued in playlist, add more!</li>
-      break
-    default:
-      queuing = (
-        <li>
-          <q>{countQueueing}</q> songs queued in playlist
-        </li>
-      )
-  }
-
-  /**
-   * Display karaoke end or playlist end
-   *
-   * The code is voluntary redundant, as the different cases do not factor
-   * well together. At least, the code is easy to understand.
-   */
-
-  const playlistEndDate =
-    dateEnd && (countQueueing || playerStatus.playlist_entry)
-      ? dayjs(dateEnd)
-      : null
-  const karaokeEndDate = karaokeDateStop ? dayjs(karaokeDateStop) : null
-
-  let end
-  // only playlist date end
-  if (playlistEndDate && !karaokeEndDate) {
-    end = (
-      <li>
-        Playlist ends at <q>{formatDate(playlistEndDate)}</q>
-      </li>
-    )
-    // only karaoke date end
-  } else if (!playlistEndDate && karaokeEndDate) {
-    if (karaokeEndDate.isAfter()) {
-      end = (
-        <>
-          <li>
-            Karaoke ends at <q>{formatDate(karaokeEndDate)}</q>
-          </li>
-          <li>
-            <q>{dayjs().to(karaokeEndDate, true)}</q> remaining
-          </li>
-        </>
-      )
-    } else {
-      end = <li>Karaoke ended</li>
-    }
-    // both playlist date end and karaoke date end
-  } else if (playlistEndDate && karaokeEndDate) {
-    // karaoke date end is after playlist date end
-    if (karaokeEndDate.isAfter(playlistEndDate)) {
-      end = (
-        <>
-          <li>
-            Karaoke ends at <q>{formatDate(karaokeEndDate)}</q>
-          </li>
-          <li>
-            <q>{dayjs().to(karaokeEndDate, true)}</q> remaining
-          </li>
-        </>
-      )
-      // playlist date end is after karaoke date end
-    } else {
-      end = (
-        <>
-          <li>
-            Playlist should end after karaoke at{' '}
-            <q>{formatDate(playlistEndDate)}</q>
-          </li>
-          <li>Playlist exceeds karaoke scheduled end!</li>
-        </>
-      )
-    }
-  }
-
-  return (
-    <CarouselEntry jumbo="Stats" className="stats">
-      <ul>
-        {played}
-        {queuing}
-        {end}
-      </ul>
-    </CarouselEntry>
-  )
-}
