+++ src/style/components/generics/_index.scss
@@ -4,11 +4,11 @@
 // Generic components main style file
 //
 
+@use 'carousel';
 @use 'details';
 @use 'form';
 @use 'listing';
 @use 'navigator';
 @use 'notification';
-@use 'playlist_entry_minimal';
 @use 'tab_bar';
 @use 'token_widget';
