--- src/style/components/generics/_playlist_entry_minimal.scss
@@ -1,11 +1,0 @@
-//
-// Dakara Project
-//
-// Minimal playlist entry widget file
-//
-
-.playlist-entry-minimal {
-  .user-widget {
-    margin-top: -0.25em;
-  }
-}
