+++ src/style/components/karaoke/_index.scss
@@ -5,4 +5,3 @@
 //
 
 @use 'player';
-@use 'player_notification';
