+++ src/style/components/karaoke/player/_index.scss
@@ -1,0 +1,9 @@
+//
+// Dakara Project
+//
+// Player main style file
+//
+
+@use 'manage_button';
+@use 'notification';
+@use 'player';
