+++ src/style/components/karaoke/player/_notification.scss
